new job system - closes #1170

+++ src/engine/mt/task.h
@@ -23,9 +23,9 @@ public:
 	bool create(const char* name);
 	bool destroy();
 
-	void setAffinityMask(u32 affinity_mask);
+	void setAffinityMask(u64 affinity_mask);
 
-	u32 getAffinityMask() const;
+	u64 getAffinityMask() const;
 
 	bool isRunning() const;
 	bool isFinished() const;
